Removed *_new files and updated the baseline
