+++ configurations/insert_only.yml
@@ -2,4 +2,4 @@ features:
   fields:
     marco:
   hcal:
-    forward_insert_new:
+    forward_insert:
